% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{blackhole}
\alias{blackhole}
\title{Black hole in Messier 87 captured by the Event Horizon Telescope}
\format{
A 250x250 matrix containing pixels evaluated between 0 and 1.
}
\source{
\url{https://commons.wikimedia.org/wiki/File:Black_hole_-_Messier_87_crop_max_res.jpg}
\url{https://mtsch.github.io/Ripserer.jl/v0.10/generated/sublevelset/}

\describe{
Data pre-processing:
First, we obtained the image from {\href{https://commons.wikimedia.org/wiki/File:Black_hole_-_Messier_87_crop_max_res.jpg}{Wikimedia Commons}}
We then utilize {\href{https://cran.r-project.org/package=magick}{{magick}}} to
convert the image from RGB to grayscale using the default
"perceptually-weighted" conversion. Next we acquired the raw 3D array,
converted the data type to numerical, and dropped the singleton channel
dimension. We then transposed the matrix and vertically flipped it to align
with how \href{https://rdocumentation.org/packages/graphics/versions/3.6.2}{{graphics}}
reads matrices.
}
}
\usage{
blackhole
}
\description{
A data set containing a grayscale bitmap of the Messier 87 black
hole with 250x250 resolution.
}
\examples{
# compute the persistence homology of the Messier 87 black hole 

# contour lines connect regions of equal brightness, revealing the structure of the glowing gas 
# surrounding the black hole
image(blackhole, col = hcl.colors(256, palette = "inferno", alpha = NULL, rev = FALSE, fixup = TRUE))
title(main = "Messier 87's Black Hole")

#' # based on the image, we expect one especially prominent persistent feature in 1D
ph <- cubical(blackhole)

plot.new()
plot.window(
  xlim = c(0, max(ph$death)),
  ylim = c(0, max(ph$death)),
  asp = 1
)
axis(1L)
axis(2L)
abline(a = 0, b = 1)
points(ph[ph$dim == 0L, c("birth", "death")], pch = 16L)
points(ph[ph$dim == 1L, c("birth", "death")], pch = 17L)
}
\keyword{datasets}
