% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{case_predictors}
\alias{case_predictors}
\title{State-level predictors of mosquito-borne illness in Brazil}
\format{
A data frame of 27 observations and 4 variables:
\describe{
\item{POP}{state population in 2013}
\item{TEMP}{average temperature across state municipalities}
\item{PRECIP}{average precipitation across state municipalities}
\item{CASE}{number of state Dengue cases in 2013}
}
}
\source{
\url{<multiple URLs>}
}
\usage{
case_predictors
}
\description{
A data set of numbers of cases of Dengue in each state of Brazil
in 2013 and three state-level variables used in a predictive model.
}
\keyword{datasets}
