% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHom.R
\name{print.PHom}
\alias{print.PHom}
\title{Printing Persistence Data}
\usage{
\method{print}{PHom}(x, ...)
}
\arguments{
\item{x}{object of class \code{PHom}}

\item{...}{other parameters; ignored}
}
\description{
Print a PHom object.
}
\examples{
# create circle dataset
angles <- runif(25, 0, 2 * pi)
circle <- cbind(cos(angles), sin(angles))

# calculate persistent homology
circle_phom <- vietoris_rips(circle)

# print persistence data
print(circle_phom)
}
