% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripser_func.R
\name{ripser}
\alias{ripser}
\title{Calculate Persistent Homology of a Point Cloud}
\usage{
ripser(mat)
}
\arguments{
\item{mat}{numeric matrix containing point cloud}
}
\description{
The `mat` parameter should be a numeric matrix with each row corresponding
to a single point, and each column corresponding to a single dimension. Thus,
if `mat` has 50 rows and 5 columns, it represents a point cloud with 50 points
in 5 dimensions.
}
