% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripserr.R
\docType{package}
\name{ripserr}
\alias{ripserr-package}
\alias{ripserr}
\title{Calculate Persistent Homology with Ripser-Based Engines}
\description{
Ports Ripser-based persistent homology calculation engines
from C++ to R using the Rcpp package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tdaverse/ripserr/}
  \item Report bugs at \url{https://github.com/tdaverse/ripserr/issues}
}

}
\author{
\strong{Maintainer}: Jason Cory Brunson \email{cornelioid@gmail.com} (\href{https://orcid.org/0000-0003-3126-9494}{ORCID})

Authors:
\itemize{
  \item Jason Cory Brunson \email{cornelioid@gmail.com} (\href{https://orcid.org/0000-0003-3126-9494}{ORCID})
  \item Raoul R. Wadhwa \email{raoulwadhwa@gmail.com} (\href{https://orcid.org/0000-0003-0503-9580}{ORCID})
  \item Matt Piekenbrock \email{matt.piekenbrock@gmail.com}
  \item Xinyi Zhang \email{ezhang0927@gmail.com}
  \item Alice Zhang \email{aliscezhang@gmail.com} (\href{https://orcid.org/0009-0001-3584-5869}{ORCID})
  \item Kent Phipps
  \item Sean Hershkowitz \email{Sea7777.hers@gmail.com}
}

Other contributors:
\itemize{
  \item Emily Noble [contributor]
  \item Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID}) [contributor]
  \item Aidan Bryant \email{bryant.aidan15@gmail.com} [contributor]
  \item James Golabek \email{jamesgolabek@gmail.com} [contributor]
  \item Jacob G. Scott (\href{https://orcid.org/0000-0003-2971-7673}{ORCID}) [laboratory director]
  \item Ulrich Bauer (Ripser; MIT license) [copyright holder, contributor]
  \item Takeki Sudo (Cubical Ripser; GPL-3 license) [copyright holder, contributor]
  \item Kazushi Ahara (Cubical Ripser; GPL-3 license) [copyright holder, contributor]
}

}
