% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{cubical}
\alias{cubical}
\title{Calculate Persistent Homology using a Cubical Complex}
\usage{
cubical(mat, threshold = 9999, method = 0)
}
\arguments{
\item{mat}{numeric matrix containing pixel data}

\item{threshold}{maximum diameter for computation of Cubical complex}
}
\value{
3-column matrix with each row representing a TDA feature
}
\description{
Temporary description.
}
