% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripser_func.R
\name{ripser}
\alias{ripser}
\title{Calculate Persistent Homology of a Point Cloud}
\usage{
ripser(mat)
}
\arguments{
\item{mat}{numeric matrix containing point cloud}
}
\value{
3-column matrix, with each row representing a TDA feature
}
\description{
The `mat` parameter should be a numeric matrix with each row corresponding
to a single point, and each column corresponding to a single dimension. Thus,
if `mat` has 50 rows and 5 columns, it represents a point cloud with 50 points
in 5 dimensions.
}
\examples{

# create a 2-d point cloud of a circle (100 points)
num.pts <- 100
rand.angle <- runif(num.pts, 0, 2*pi)
pt.cloud <- cbind(cos(rand.angle), sin(rand.angle))

# calculate persistent homology (num.pts by 3 numeric matrix)
pers.hom <- ripser(pt.cloud)
}
