% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripserr.R
\docType{package}
\name{ripserr}
\alias{ripserr}
\title{Calculate Persistent Homology with Ripser-Based Engines}
\description{
Ports Ripser-based persistent homology calculation engines
from C++ to R using the Rcpp package.
}
