% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vietoris_rips.R
\name{vietoris_rips}
\alias{vietoris_rips}
\alias{vietoris_rips.data.frame}
\alias{vietoris_rips.matrix}
\alias{vietoris_rips.dist}
\alias{vietoris_rips.numeric}
\alias{vietoris_rips.ts}
\alias{vietoris_rips.default}
\title{Calculate Persistent Homology via a Vietoris-Rips Complex}
\usage{
vietoris_rips(dataset, ...)

\method{vietoris_rips}{data.frame}(dataset, ...)

\method{vietoris_rips}{matrix}(dataset, max_dim = 1L, threshold = -1, p = 2L, ...)

\method{vietoris_rips}{dist}(dataset, max_dim = 1L, threshold = -1, p = 2L, ...)

\method{vietoris_rips}{numeric}(
  dataset,
  data_dim = 2L,
  dim_lag = 1L,
  sample_lag = 1L,
  method = "qa",
  ...
)

\method{vietoris_rips}{ts}(dataset, ...)

\method{vietoris_rips}{default}(dataset, ...)
}
\description{
Calculate Persistent Homology via a Vietoris-Rips Complex

Vietoris-Rips Persistent Homology of a Data Frame

Vietoris-Rips Persistent Homology of a Matrix

Vietoris-Rips Persistent Homology of a \code{dist} Object

Vietoris-Rips Persistent Homology of Time Series

Default Vietoris-Rips Persistent Homology
}
